% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBN_estim.R
\name{cached_local_score}
\alias{cached_local_score}
\title{main function computing the local score i.e. the regression on a specific gene
this function use the global variable "global" which must be created by the compute_global function
the function check if the result is stored in global$cache and return or compute it}
\usage{
cached_local_score(j, dag, crit = c("eBIC", "BIC"))
}
\arguments{
\item{j}{one gene index}

\item{dag}{a dag object (we only need the parents of j, but it is simpler to pass the whole dag)}

\item{crit}{either the extended BIC (default) or the BIC. The extended BIC should select more parsimonious models}
}
\value{
the local score which is the likelihood of the j component of the model plus the penalty
}
\description{
main function computing the local score i.e. the regression on a specific gene
this function use the global variable "global" which must be created by the compute_global function
the function check if the result is stored in global$cache and return or compute it
}
