% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBN_estim.R
\name{compute_global}
\alias{compute_global}
\title{function creating the workspace for fitting our models
the output must be named "global" and used as a global variable}
\usage{
compute_global(x, N, I)
}
\arguments{
\item{x}{the (n,p) real value (n experiments, p genes)}

\item{N}{vector of length k, number of replication per design}

\item{I}{list of length k, intervention pattern per design (\code{NULL} means "no intervention")
e.g. \code{N=c(20,5,5)} and \code{I=list(NULL,c(1,3),c(2,3))} => 20 WT + 5 KO on 1 and 3 + 5 KO on 2 and 3}
}
\value{
a list containing
- x the data, p number of genes, n number of experiments
- idx a list of length p such that \code{idx[[j]]} contains all experiments with no intervention on j
- cache is a empty hashmap with numeric values, further used for speeding up computations
}
\description{
function creating the workspace for fitting our models
the output must be named "global" and used as a global variable
}
