% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_utils.R
\name{dag2dot}
\alias{dag2dot}
\title{convert a dag object into dot file}
\usage{
dag2dot(dag, file = "", w = NULL)
}
\arguments{
\item{dag}{a dag object}

\item{file}{a file for the output}

\item{w}{optional weights for the edges}
}
\value{
a dot description of the dag
}
\description{
convert a dag object into dot file
}
\examples{
dag2dot(bn(c(),c(1),c(1,2)))

}
