% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBN_utils.R
\name{dag2ref}
\alias{dag2ref}
\title{Generates a reference GBN from a DAG}
\usage{
dag2ref(
  dag,
  w.range = c(-3, 3),
  m.range = c(-2, 2),
  s.mean = log(1),
  s.sd = 0.3
)
}
\arguments{
\item{dag}{a DAG object}

\item{w.range}{range for the direct effect (uniform distribution)}

\item{m.range}{range for the residual means (uniform distribution)}

\item{s.mean}{mean of the log of residual}

\item{s.sd}{sd of the log of residual}
}
\value{
a GBN
}
\description{
Generates a reference GBN from a DAG
}
\examples{
dag=str2dag("[][1][1,2]")
dag2ref(dag)

}
