% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_utils.R
\name{dag2str}
\alias{dag2str}
\title{convert a dag object into a string}
\usage{
dag2str(dag)
}
\arguments{
\item{dag}{a dag object}
}
\value{
a string
}
\description{
convert a dag object into a string
}
\examples{
dag2str(bn(c(),c(1),c(1,2)))

}
