% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBN_estim.R
\name{estim_gbn}
\alias{estim_gbn}
\title{simple function estimating the model for a given dag
this function use the global variable "global" which must be created by the compute_global function}
\usage{
estim_gbn(dag)
}
\arguments{
\item{dag}{a DAG}
}
\value{
an estimated GBN
}
\description{
simple function estimating the model for a given dag
this function use the global variable "global" which must be created by the compute_global function
}
