% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBN_estim.R
\name{mc3_gbn}
\alias{mc3_gbn}
\title{The MC3 implementation using the cache mecanism
this function use the global variable "global" which must be created by the compute_global function}
\usage{
mc3_gbn(
  itermax = 5000,
  burnin = 1000,
  initial = NULL,
  maxParents = global$p - 1,
  constraintT = matrix(0, global$p, global$p),
  verbose = TRUE,
  exact = TRUE,
  crit = c("eBIC", "BIC")
)
}
\arguments{
\item{itermax}{max number of iterations}

\item{burnin}{the number of iterations to discard at the end}

\item{initial}{a starting DAG or NULL for random start}

\item{maxParents}{the maximum number of parents for each node}

\item{constraintT}{a matrix of additional constraints (0: no constraint, 1: edge must be present, -1: must be absent)}

\item{verbose}{boolean for verbose output}

\item{exact}{boolean for exact posterior on the approximate support}

\item{crit}{criterion for the score}
}
\value{
the MC3 GBN estimate
}
\description{
The MC3 implementation using the cache mecanism
this function use the global variable "global" which must be created by the compute_global function
}
