% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBN_utils.R
\name{simul_gbn}
\alias{simul_gbn}
\title{Simulate a GBN with a specific intervention design}
\usage{
simul_gbn(
  m,
  sigma,
  W,
  N,
  I = list(c()),
  fill_int = function(n) {     rep(0, n) }
)
}
\arguments{
\item{m}{vector of residual means}

\item{sigma}{vector of residual standard deviations}

\item{W}{matrix of direct effects}

\item{N}{either an integer or a vector of integers}

\item{I}{a list of intervention sets}

\item{fill_int}{a function of \code{n} returning a vector of \code{n} data to fill the intervention genes}
}
\value{
the simulated data matrix with \code{sum(N)} rows
}
\description{
Simulate a GBN with a specific intervention design
}
\examples{
dag=bn(c(),1,2)
set.seed(42)
ref=dag2ref(dag)
N=c(50,5,5)
I=list(numeric(0),c(1,3),2)
x=simul_gbn(ref$m,ref$sigma,ref$W,N,I)
x=simul_gbn(ref$m,ref$sigma,ref$W,N,I,fill_in=function(n)rnorm(n))

}
