% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_utils.R
\name{str2dag}
\alias{str2dag}
\title{convert a string into a dag object}
\usage{
str2dag(str)
}
\arguments{
\item{str}{a string (ex: \code{"[][2][3]"})}
}
\value{
a dag object
}
\description{
convert a string into a dag object
}
\examples{
str2dag("[][1][1,2]")


}
